src/common/guards/firebase-auth.guard.ts
import { Injectable, SetMetadata, ExecutionContext } from '@nestjs/common';
import { AuthGuard } from '@nestjs/passport';
import { Reflector } from '@nestjs/core';
import { Observable } from 'rxjs';

export const IS_PUBLIC__METADATA_KEY = Symbol('isPublic');
export const Public = () => SetMetadata(IS_PUBLIC__METADATA_KEY, true);

@Injectable()
export class FirebaseAuthGuard extends AuthGuard('bearer') {
  constructor(private readonly reflector: Reflector) {
    super();
  }

  override canActivate(
    context: ExecutionContext,
  ): boolean | Promise<boolean> | Observable<boolean> {
    const isPublic = this.reflector.getAllAndOverride<boolean>(
      IS_PUBLIC__METADATA_KEY,
      [context.getHandler(), context.getClass()],
    );

    if (isPublic) {
      return true;
    }

    return super.canActivate(context);
  }
}
